+++ dwarfeng-dct-core/src/main/java/com/dwarfeng/dct/stack/util/ValueCodingConfigUtil.java
@@ -1,5 +1,6 @@
-package com.dwarfeng.dct.sdk.util;
+package com.dwarfeng.dct.stack.util;
 
+import com.dwarfeng.dct.sdk.util.Constants;
 import com.dwarfeng.dct.stack.handler.ValueCodec;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Strings;
@@ -8,12 +9,12 @@ import java.util.Collection;
 import java.util.Objects;
 
 /**
- * 值编解码器工具类。
+ * 值编解码配置工具类。
  *
  * @author DwArFeng
- * @since 1.0.0
+ * @since 3.0.0
  */
-public final class ValueCodecUtil {
+public final class ValueCodingConfigUtil {
 
     /**
      * 检查指定的值编解码器是否合法。
@@ -22,7 +23,7 @@ public final class ValueCodecUtil {
      */
     public static void checkCodec(ValueCodec valueCodec) {
         if (Objects.isNull(valueCodec)) {
-            throw new IllegalArgumentException("ValueCodec 不能为空");
+            throw new NullPointerException("值编解码器不能为 null");
         }
         checkTargetClass(valueCodec.getTargetClass());
         checkValuePrefix(valueCodec.getValuePrefix());
@@ -35,7 +36,7 @@ public final class ValueCodecUtil {
      */
     public static void checkCodecs(Collection<ValueCodec> valueCodecs) {
         if (Objects.isNull(valueCodecs)) {
-            throw new IllegalArgumentException("ValueCodec 列表不能为空");
+            throw new NullPointerException("值编解码器列表不能为 null");
         }
         for (ValueCodec valueCodec : valueCodecs) {
             checkCodec(valueCodec);
@@ -49,7 +50,7 @@ public final class ValueCodecUtil {
      */
     public static void checkTargetClass(Class<?> targetClass) {
         if (Objects.isNull(targetClass)) {
-            throw new IllegalArgumentException("目标类不能为空");
+            throw new NullPointerException("目标类不能为 null");
         }
     }
 
@@ -60,7 +61,7 @@ public final class ValueCodecUtil {
      */
     public static void checkTargetClasses(Collection<Class<?>> targetClasses) {
         if (Objects.isNull(targetClasses)) {
-            throw new IllegalArgumentException("目标类列表不能为空");
+            throw new NullPointerException("目标类列表不能为 null");
         }
         for (Class<?> targetClass : targetClasses) {
             checkTargetClass(targetClass);
@@ -74,7 +75,7 @@ public final class ValueCodecUtil {
      */
     public static void checkValuePrefix(String valuePrefix) {
         if (Objects.isNull(valuePrefix)) {
-            throw new IllegalArgumentException("值前缀不能为空");
+            throw new NullPointerException("值前缀不能为 null");
         }
         if (StringUtils.isEmpty(valuePrefix)) {
             throw new IllegalArgumentException("值前缀不能为空");
@@ -94,14 +95,14 @@ public final class ValueCodecUtil {
      */
     public static void checkValuePrefixes(Collection<String> valuePrefixes) {
         if (Objects.isNull(valuePrefixes)) {
-            throw new IllegalArgumentException("值前缀列表不能为空");
+            throw new NullPointerException("值前缀列表不能为 null");
         }
         for (String valuePrefix : valuePrefixes) {
             checkValuePrefix(valuePrefix);
         }
     }
 
-    private ValueCodecUtil() {
+    private ValueCodingConfigUtil() {
         throw new IllegalStateException("禁止外部实例化");
     }
 }